Standardize capitalization
I've attempted to standardize capitalization and punctuation, both so the site looks better, and so that our translations are more uniform.

Basically, any small prompts or short text that would be found on a button does not end on a period. Any longer text, particularly information that would appear on a textbox, had punctuation added. Capital letters are only used for the first word in a sentence or prompt.

Further discussion on what things to capitalize/punctuate or not is welcome.