Translation fixes to include artist name
* added a single point

at the end of a sentence.
some could say it was... pointless (but hopefully not this commit)

* change "the/this artist" to variable with artist name + some wording

in cases that seemed relevant, changed "the/this artist" to something to a varaible for the actual artist name.

* added artistName variable

* added artistName variable

* added artistName variable

* added artistName variable

* added artistName variable

* added artistName variable

* added artistName variable

* added artistName variable

* added artistName variable

not sure about this one, the syntax highlighter doesn't seem to like what i did

* added artistName variable to one string

mirroring its twin "stopReceivingEmails"

---------

Co-authored-by: Cwpute <80604512+Cwpute@users.noreply.github.com>