Updating README.md
started work to improve the Readme.

Not sure what to put in the Features section, since it will certainly be a mix of backend stuff and user-facing stuff. Same with the "How it works" section as i don't know, well, how it works :upside_down_face: 

Also changed the sections organisation after "Contributing" as it seemed incorrect.